requirements.txt
gradio==5.35.0
tensorflow
numpy
pillow
